Add ability to fix zip corruption (#87)
`@turbowarp/jszip@3.12.0` now includes a recovery mode where it tries to find file entries by scanning the zip if the central directory is missing. The way that JSZip is really nice as it puts all the sizes and everything directly in the local header, so we don't need to do a lot of guesswork. If we had to support a bunch of random zip generators, this would be more work.

The sb3fix changes are to enable this mode, add logs, and remove files that have mismatched uncompressed data sizes so that regular Scratch will be able to load it.

These projects now get automatically fixed instead of waiting for me to run `zip -FF` manually:

 - All 3 projects in https://github.com/TurboWarp/sb3fix/issues/82#issuecomment-3418751929
 - https://github.com/TurboWarp/sb3fix/issues/80
 - https://github.com/TurboWarp/sb3fix/issues/59
 - https://github.com/TurboWarp/sb3fix/issues/32
 - https://github.com/TurboWarp/sb3fix/issues/31
 - https://github.com/TurboWarp/sb3fix/issues/30
 - https://github.com/TurboWarp/sb3fix/issues/28
 - https://github.com/TurboWarp/sb3fix/issues/7

It probably would've automatically fixed these but the links no longer work, so can't verify for certain:

 - https://github.com/TurboWarp/sb3fix/issues/81
 - https://github.com/TurboWarp/sb3fix/issues/50
 - https://github.com/TurboWarp/sb3fix/issues/22
 - https://github.com/TurboWarp/sb3fix/issues/5
 - https://github.com/TurboWarp/sb3fix/issues/4
 - https://github.com/TurboWarp/sb3fix/issues/3